requirements.txt
langchain
langchain-openai
langchain-community
chromadb
pypdf
python-dotenv
tiktoken